@@ -3,19 +3,20 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.0
-  provider: ^4.3.2+2
-  google_fonts: ^1.1.2
+  cupertino_icons: ^1.0.3
+  provider: ^5.0.0
+  google_fonts: ^2.1.0
 
 
 dev_dependencies:
-  integration_test: ^1.0.0
-  test: ^1.9.4
+  integration_test:
+    sdk: flutter
+  test: ^1.16.8
   flutter_test:
     sdk: flutter
 
