@@ -1,5 +1,3 @@
-import 'package:flutter/material.dart';
-
 enum TransactionType { expense, income }
 
 extension XExpenseType on TransactionType {
@@ -13,13 +11,10 @@ class Transaction {
   final double amount;
   final TransactionType type;
 
-  Transaction({
-    @required this.description,
-    @required this.date,
-    @required this.amount,
-    @required this.type,
-  })  : assert(description != null),
-        assert(date != null),
-        assert(type != null),
-        assert(amount != null);
+  const Transaction({
+    required this.description,
+    required this.date,
+    required this.amount,
+    required this.type,
+  });
 }
