@@ -1,21 +1,19 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_transaction_tracker/transaction_screen.dart';
-import 'package:flutter_transaction_tracker/transaction_provider.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
+import 'transaction_provider.dart';
+import 'transaction_screen.dart';
 
 void main() {
   runApp(const TransactionTrackerApp());
 }
 
 class TransactionTrackerApp extends StatelessWidget {
-  const TransactionTrackerApp({Key key}) : super(key: key);
+  const TransactionTrackerApp({Key? key}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'TransactionTrackerApp',
-      // This will force Flutter to generate
-      //  an overlay to visualize the Semantics tree.
       showSemanticsDebugger: true,
       theme: ThemeData(
         primarySwatch: Colors.green,
