@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_transaction_tracker/domain/transaction.dart';
-import 'package:flutter_transaction_tracker/transaction_provider.dart';
-import 'package:flutter_transaction_tracker/transaction_item.dart';
-import 'package:flutter_transaction_tracker/transaction_modal.dart';
 import 'package:provider/provider.dart';
+import 'domain/transaction.dart';
+import 'transaction_item.dart';
+import 'transaction_modal.dart';
+import 'transaction_provider.dart';
 
 class TransactionScreen extends StatelessWidget {
-  const TransactionScreen({Key key}) : super(key: key);
+  const TransactionScreen({Key? key}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     final provider = context.watch<TransactionProvider>()
@@ -39,7 +39,7 @@ class TransactionScreen extends StatelessWidget {
 
 class BalanceAmount extends StatelessWidget {
   final double amount;
-  const BalanceAmount({Key key, this.amount}) : super(key: key);
+  const BalanceAmount({Key? key, required this.amount}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     return Center(
@@ -67,7 +67,8 @@ class BalanceAmount extends StatelessWidget {
 
 class TransactionListBody extends StatelessWidget {
   final List<Transaction> transactions;
-  const TransactionListBody({Key key, this.transactions}) : super(key: key);
+  const TransactionListBody({Key? key, required this.transactions})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
