@@ -1,2 +1,1 @@
-bool isNumeric(String input) =>
-    RegExp(r'^-?(([0-9]*)|(([0-9]*)\.([0-9]*)))$').hasMatch(input);
+bool isNumeric(String input) => RegExp(r'^-?(([0-9]*)|(([0-9]*)\.([0-9]*)))$').hasMatch(input);
